move curves functions to manager
